Create encrypted shared prefs only once.
Creating the shared prefs instance is an operation that isn't trivial
(takes ~50ms on a OnePlus 6), so avoid doing it over and over again.
