+++ mobile/src/main/java/org/openhab/habdroid/util/ExtensionFuncs.kt
@@ -30,9 +30,6 @@ import android.util.Log
 import androidx.annotation.StringRes
 import androidx.core.content.ContextCompat
 import androidx.preference.PreferenceManager
-import android.os.Build
-import androidx.security.crypto.EncryptedSharedPreferences
-import androidx.security.crypto.MasterKeys
 import com.caverock.androidsvg.SVG
 import com.caverock.androidsvg.SVGParseException
 import es.dmoral.toasty.Toasty
@@ -41,6 +38,7 @@ import okhttp3.ResponseBody
 import org.json.JSONArray
 import org.json.JSONObject
 import org.openhab.habdroid.R
+import org.openhab.habdroid.core.OpenHabApplication
 import org.w3c.dom.Node
 import org.w3c.dom.NodeList
 import java.io.IOException
@@ -204,6 +202,10 @@ fun Context.getPrefs(): SharedPreferences {
     return PreferenceManager.getDefaultSharedPreferences(this)
 }
 
+fun Context.getSecretPrefs(): SharedPreferences {
+    return (applicationContext as OpenHabApplication).secretPrefs
+}
+
 /**
  * Shows an orange Toast with the openHAB icon. Can be called from the background.
  */
@@ -223,19 +225,4 @@ fun Context.showToast(@StringRes message: Int) {
 
 fun Context.hasPermission(permission: String): Boolean {
     return ContextCompat.checkSelfPermission(this, permission) != PackageManager.PERMISSION_GRANTED
-}
-
-fun Context.getSecretPrefs(): SharedPreferences {
-    return if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) {
-        val masterKeyAlias = MasterKeys.getOrCreate(MasterKeys.AES256_GCM_SPEC)
-        EncryptedSharedPreferences.create(
-            "secret_shared_prefs_encrypted",
-            masterKeyAlias,
-            this,
-            EncryptedSharedPreferences.PrefKeyEncryptionScheme.AES256_SIV,
-            EncryptedSharedPreferences.PrefValueEncryptionScheme.AES256_GCM
-        )
-    } else {
-        getSharedPreferences("secret_shared_prefs", Context.MODE_PRIVATE)
-    }
 }
